# Conflicts:
#	README.md
#	examples/opensora_hpcai/scripts/inference_vae.py
#	mindone/trainers/zero.py

+++ mindone/diffusers/pipelines/controlnet/pipeline_controlnet_blip_diffusion.py
@@ -169,7 +169,7 @@ class BlipDiffusionControlNetPipeline(DiffusionPipeline):
             latents = latents.to(dtype=dtype)
 
         # scale the initial noise by the standard deviation required by the scheduler
-        latents = latents * self.scheduler.init_noise_sigma
+        latents = (latents * self.scheduler.init_noise_sigma).to(dtype)
         return latents
 
     def encode_prompt(self, query_embeds, prompt):
